requirements.txt
spacy
pytesseract
opencv-python
pillow
PyMuPDF
